Orden 6

+++ compras/views.py
@@ -12,7 +12,7 @@ from compras.forms import OrdenCompraCreateForm, OrdenCompraEditForm, DetalleOrd
 # Forms import<--
 
 # Utils import-->
-from .utils import save_detalleorden
+from .utils import fill_data, recalcular_total_orden
 # Utils import<--
 
 # Extra python features-->
@@ -96,8 +96,8 @@ class OrdenEditView(BasicEMixin, TemplateView):
         detalle = DetalleOrdenCompra.objects.filter(ordencompra=self.kwargs['pk'])
         content_detalle = []
         for idx, d in enumerate(detalle):
-            content_detalle.append([DetalleOrdenCompraForm(instance=d, proveedor=orden.proveedor_id,
-                                                           has_data=True, prefix='do'+str(idx+1)), d])
+            content_detalle.append([DetalleOrdenCompraForm(instance=d, prefix='do'+str(idx+1),
+                                                           proveedor=orden.proveedor_id, has_data=True), d])
         context['detalle'] = content_detalle
         context['clean_form'] = DetalleOrdenCompraForm(proveedor=orden.proveedor_id, has_data=False)
         return context
@@ -106,19 +106,24 @@ class OrdenEditView(BasicEMixin, TemplateView):
         orden = OrdenCompra.objects.get(pk=self.kwargs['pk'])
         form = OrdenCompraEditForm(request.POST, instance=orden)
         if form.is_valid():
-            orden = form.save()
-            for i in request.POST['detalleorden_to_save']:
-                if 'do'+i+'-id' in self.request.POST:
-                    do = DetalleOrdenCompra.objects.get(pk=self.request.POST['do'+i+'-id'])
-                    do_form = DetalleOrdenCompraForm(instance=do, prefix='do'+i)
-                else:
-                    do_form = DetalleOrdenCompraForm(prefix='do'+i)
-                do_obj = do_form.save(commit=False)
-                do_obj.ordencompra = orden
-                print(do_obj)
-                save_detalleorden(do_obj)
-            for j in request.POST['detalleorden_to_delete']:
-                DetalleOrdenCompra.objects.get(pk=j).delete()
+            orden = form.save(commit=False)
+            if request.POST['detalleorden_to_save'] != '':
+                for i in request.POST['detalleorden_to_save'].split(','):
+                    if 'do'+i+'-id' in self.request.POST:
+                        do = DetalleOrdenCompra.objects.get(pk=self.request.POST['do'+i+'-id'])
+                        do_form = DetalleOrdenCompraForm(request.POST, instance=do, prefix='do'+i,
+                                                         proveedor=orden.proveedor_id, has_data=True)
+                    else:
+                        do_form = DetalleOrdenCompraForm(request.POST, prefix='do'+i,
+                                                         proveedor=orden.proveedor_id, has_data=True)
+                    if do_form.is_valid():
+                        do_obj = do_form.save(commit=False)
+                        do_obj.ordencompra = orden
+                        fill_data(do_obj)
+            if request.POST['detalleorden_to_delete'] != '':
+                for j in request.POST['detalleorden_to_delete'].split(','):
+                    DetalleOrdenCompra.objects.get(pk=j).delete()
+            recalcular_total_orden(orden)
         else:
             return HttpResponse(form.errors)
-        return redirect('/compras/orden')
+        return redirect('/compras/orden/'+str(orden.id))
